safewebp-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pillow-heif==0.13.0
aiofiles==23.2.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
numpy==1.26.2
joblib==1.3.2